Fix posting to work again
Posting was broken (would silently fail) if an id was already taken.  So fixed Flaps and Users to have the database assign identifiers instead.

This is a bit of a footgun, since the typesystem doesnt' know that an id isn't a valid one when creating a Flap or User, but I can fix that in a later pass.

Added a timeout (5s) to all posts from the client.  That eventually should be adjustable instead.

Added an always-visible timeline refresh button in lieu of making pull-to-refresh work yet.

Made the retry dialog actually do something.