Move to mongodb instead of sembast
This appear to work locally.  Turns out the "json" mongo_dart expects is really json++, so I had to split out the network json from the dbJson to have ObjectIds in the dbJson but Strings in the network json.

The database now expects a env.json file with a MONGO_URL db url in it.  I'm not sure this is necessary, since the db access is also locked down in other ways, but using a secrets file for now.