Remove mongo_dart dependency
Moves DataStore into shorebird and makes generic.  Also hides the fact that it's implemented on top of mongodb (for now) so it can be changed later as needed (or an implementation written on top of sembast again to allow running on device).

This adds the concept of ClassInfo which is metadata about classes which is passed to the DataSTore so they can be constructed directly instead of needing to hand json to the DataStore.